+++ docs/release_notes.rst
@@ -4,11 +4,14 @@ Release notes
 =============
 
 
+* Merging in pull request https://github.com/lucabaldini/ixpeobssim/pull/657
 * Merging in pull request https://github.com/lucabaldini/ixpeobssim/pull/659
 * Merging in pull request https://github.com/lucabaldini/ixpeobssim/pull/660
 * Fix in the error propagation in the histogram sum
 * Better handling of the parameter bounds when summing fitting models.
+* New interface to level-1 files and single-event display.
 * Issue(s) closed:
+      * https://github.com/lucabaldini/ixpeobssim/issues/639
       * https://github.com/lucabaldini/ixpeobssim/issues/656
       * https://github.com/lucabaldini/ixpeobssim/issues/658